+++ experiments/mixformer_cvt/baseline.yaml
@@ -1,7 +1,5 @@
 DATA:
-  SAMPLER_MODE: trident_pro
-  MAX_SAMPLE_INTERVAL:
-  - 200
+  MAX_SAMPLE_INTERVAL: 200
   MEAN:
   - 0.485
   - 0.456
@@ -23,8 +21,14 @@ DATA:
     NUMBER: 2
   TRAIN:
     DATASETS_NAME:
-    - GOT10K_train_full
+    - GOT10K_vottrain
+    - LASOT
+    - COCO17
+    - TRACKINGNET
     DATASETS_RATIO:
+    - 1
+    - 1
+    - 1
     - 1
     SAMPLE_PER_EPOCH: 60000
   VAL:
@@ -34,9 +38,10 @@ DATA:
     - 1
     SAMPLE_PER_EPOCH: 10000
 MODEL:
+  VIT_TYPE: 'cvt_21'
   BACKBONE:
-    PRETRAINED: false
-    PRETRAINED_PATH: ''
+    PRETRAINED: True
+    PRETRAINED_PATH: '/YOUR/PRETRAINED/CVT/DIR/CvT-21-384x384-IN-22k.pth' #'/home/cyt/project/CvtT/models/CvT-21-384x384-IN-22k.pth'
     INIT: 'trunc_norm'
     NUM_STAGES: 3
     PATCH_SIZE: [ 7, 3, 3 ]
@@ -59,26 +64,21 @@ MODEL:
     PADDING_Q: [ 1, 1, 1 ]
     STRIDE_Q: [ 1, 1, 1 ]
     FREEZE_BN: true
-  HEAD_FREEZE_BN: True
-  PRETRAINED_STAGE1: True
-  HEAD_TYPE: CORNER #CORNER
+  HEAD_TYPE: CORNER
   HIDDEN_DIM: 384
-  NLAYER_HEAD: 3
-  NUM_OBJECT_QUERIES: 1
+  NUM_OBJECT_QUERIES: 1 # use when tokenHead is applied.
   POSITION_EMBEDDING: sine
   PREDICT_MASK: false
 TRAIN:
   BACKBONE_MULTIPLIER: 0.1
-  BATCH_SIZE: 32
+  BATCH_SIZE: 8  # 8 for 2080ti (maybe 10), 32 for tesla V100(32 G)
   DEEP_SUPERVISION: false
-  EPOCH: 40
-  TRAIN_SCORE: true
-  GIOU_WEIGHT: 2.0
+  EPOCH: 500
+  IOU_WEIGHT: 2.0
   GRAD_CLIP_NORM: 0.1
   L1_WEIGHT: 5.0
-  SCORE_WEIGHT: 1.0
-  LR: 0.0001 #0.0001
-  LR_DROP_EPOCH: 30
+  LR: 0.0001
+  LR_DROP_EPOCH: 400
   NUM_WORKER: 8
   OPTIMIZER: ADAMW
   PRINT_INTERVAL: 50
@@ -88,22 +88,14 @@ TRAIN:
   VAL_EPOCH_INTERVAL: 5
   WEIGHT_DECAY: 0.0001
 TEST:
-  EPOCH: 40
+  EPOCH: 500
   SEARCH_FACTOR: 5.0
   SEARCH_SIZE: 320
   TEMPLATE_FACTOR: 2.0
   TEMPLATE_SIZE: 128
   UPDATE_INTERVALS:
     LASOT: [200]
-    GOT10K_TEST: [25]
+    GOT10K_TEST: [200]
     TRACKINGNET: [25]
     VOT20: [10]
     VOT20LT: [200]
-  ONLINE_SIZES:
-    LASOT: [3]
-    UAV: [3]
-    OTB: [3]
-    GOT10K_TEST: [3]
-    TRACKINGNET: [3]
-    VOT20: [3]
-    VOT20LT: [3]
