+++ experiments/mixformer_cvt/baseline_1k.yaml
@@ -1,7 +1,5 @@
 DATA:
-  SAMPLER_MODE: trident_pro
-  MAX_SAMPLE_INTERVAL:
-  - 200
+  MAX_SAMPLE_INTERVAL: 200
   MEAN:
   - 0.485
   - 0.456
@@ -40,21 +38,22 @@ DATA:
     - 1
     SAMPLE_PER_EPOCH: 10000
 MODEL:
+  VIT_TYPE: 'cvt_21'
   BACKBONE:
-    PRETRAINED: false
-    PRETRAINED_PATH: ''
+    PRETRAINED: True
+    PRETRAINED_PATH: '/YOUR/PRETRAINED/CVT/DIR/CvT-21-384x384-IN-1k.pth'    #'/home/cyt/project/CvtT/models/CvT-21-384x384-IN-1k.pth'
     INIT: 'trunc_norm'
     NUM_STAGES: 3
     PATCH_SIZE: [ 7, 3, 3 ]
     PATCH_STRIDE: [ 4, 2, 2 ]
     PATCH_PADDING: [ 2, 1, 1 ]
-    DIM_EMBED: [ 192, 768, 1024 ]
-    NUM_HEADS: [ 3, 12, 16 ]
-    DEPTH: [ 2, 2, 12 ]
+    DIM_EMBED: [ 64, 192, 384 ]
+    NUM_HEADS: [ 1, 3, 6 ]
+    DEPTH: [ 1, 4, 16 ]
     MLP_RATIO: [ 4.0, 4.0, 4.0 ]
     ATTN_DROP_RATE: [ 0.0, 0.0, 0.0 ]
     DROP_RATE: [ 0.0, 0.0, 0.0 ]
-    DROP_PATH_RATE: [ 0.0, 0.0, 0.3 ]
+    DROP_PATH_RATE: [ 0.0, 0.0, 0.1 ]
     QKV_BIAS: [ True, True, True ]
     CLS_TOKEN: [ False, False, False ]
     POS_EMBED: [ False, False, False ]
@@ -65,27 +64,21 @@ MODEL:
     PADDING_Q: [ 1, 1, 1 ]
     STRIDE_Q: [ 1, 1, 1 ]
     FREEZE_BN: true
-  HEAD_FREEZE_BN: True
-  PRETRAINED_STAGE1: True
-#  STAGE1_MODEL: '/data/cyt/experiments/cvtt/models/TSP_Online_cvtw16_22k_ep0475.pth.tar'
-  HEAD_TYPE: CORNER
-  HIDDEN_DIM: 1024
-  NLAYER_HEAD: 3
+  HEAD_TYPE: CORNER #CORNER
+  HIDDEN_DIM: 384
   NUM_OBJECT_QUERIES: 1
   POSITION_EMBEDDING: sine
   PREDICT_MASK: false
 TRAIN:
-  BACKBONE_MULTIPLIER: 0.1
-  BATCH_SIZE: 16
+  BACKBONE_MULTIPLIER: 1.0
+  BATCH_SIZE: 8
   DEEP_SUPERVISION: false
-  EPOCH: 40
-  TRAIN_SCORE: true
-  GIOU_WEIGHT: 2.0
+  EPOCH: 500
+  IOU_WEIGHT: 2.0
   GRAD_CLIP_NORM: 0.1
   L1_WEIGHT: 5.0
-  SCORE_WEIGHT: 1.0
-  LR: 0.0001
-  LR_DROP_EPOCH: 30
+  LR: 0.0001 #0.0001
+  LR_DROP_EPOCH: 400
   NUM_WORKER: 8
   OPTIMIZER: ADAMW
   PRINT_INTERVAL: 50
@@ -95,24 +88,14 @@ TRAIN:
   VAL_EPOCH_INTERVAL: 5
   WEIGHT_DECAY: 0.0001
 TEST:
-  EPOCH: 40
+  EPOCH: 500
   SEARCH_FACTOR: 5.0
   SEARCH_SIZE: 320
   TEMPLATE_FACTOR: 2.0
   TEMPLATE_SIZE: 128
   UPDATE_INTERVALS:
-    LASOT: [ 200 ]
-    UAV: [ 25 ]
-    OTB: [ 6 ]
-    GOT10K_TEST: [ 200 ]
-    TRACKINGNET: [ 25 ]
-    VOT20: [ 10 ]
-#    VOT20LT: [ 200 ]
-  ONLINE_SIZES:
-    LASOT: [ 2 ]
-    UAV: [ 3 ]
-    OTB: [ 2 ]
-    GOT10K_TEST: [ 2 ]
-    TRACKINGNET: [ 3 ]
-    VOT20: [ 5 ]
-#    VOT20LT: [ 3 ]
+    LASOT: [200]
+    GOT10K_TEST: [200]
+    TRACKINGNET: [25]
+    VOT20: [10]
+    VOT20LT: [200]
