+++ experiments/mixformer_cvt/baseline_got.yaml
@@ -32,6 +32,7 @@ DATA:
     - 1
     SAMPLE_PER_EPOCH: 10000
 MODEL:
+  VIT_TYPE: 'cvt_21'
   BACKBONE:
     PRETRAINED: True
     PRETRAINED_PATH: '/YOUR/PRETRAINED/CVT/DIR/CvT-21-384x384-IN-22k.pth' #'/home/cyt/project/CvtT/models/CvT-21-384x384-IN-22k.pth'
@@ -67,7 +68,7 @@ TRAIN:
   BATCH_SIZE: 32
   DEEP_SUPERVISION: false
   EPOCH: 500
-  GIOU_WEIGHT: 2.0
+  IOU_WEIGHT: 2.0
   GRAD_CLIP_NORM: 0.1
   L1_WEIGHT: 5.0
   LR: 0.0001
