+++ experiments/mixformer_cvt/baseline_large.yaml
@@ -38,21 +38,22 @@ DATA:
     - 1
     SAMPLE_PER_EPOCH: 10000
 MODEL:
+  VIT_TYPE: 'cvt_24'
   BACKBONE:
     PRETRAINED: True
-    PRETRAINED_PATH: '/YOUR/PRETRAINED/CVT/DIR/CvT-21-384x384-IN-1k.pth'    #'/home/cyt/project/CvtT/models/CvT-21-384x384-IN-1k.pth'
+    PRETRAINED_PATH: '/YOUR/PRETRAINED/CVT/DIR/CvT-w24-384x384-IN-22k.pth' #'/home/cyt/project/CvtT/models/CvT-w24-384x384-IN-22k.pth'
     INIT: 'trunc_norm'
     NUM_STAGES: 3
     PATCH_SIZE: [ 7, 3, 3 ]
     PATCH_STRIDE: [ 4, 2, 2 ]
     PATCH_PADDING: [ 2, 1, 1 ]
-    DIM_EMBED: [ 64, 192, 384 ]
-    NUM_HEADS: [ 1, 3, 6 ]
-    DEPTH: [ 1, 4, 16 ]
+    DIM_EMBED: [ 192, 768, 1024 ]
+    NUM_HEADS: [ 3, 12, 16 ]
+    DEPTH: [ 2, 2, 12 ]
     MLP_RATIO: [ 4.0, 4.0, 4.0 ]
     ATTN_DROP_RATE: [ 0.0, 0.0, 0.0 ]
     DROP_RATE: [ 0.0, 0.0, 0.0 ]
-    DROP_PATH_RATE: [ 0.0, 0.0, 0.1 ]
+    DROP_PATH_RATE: [ 0.0, 0.0, 0.3 ]
     QKV_BIAS: [ True, True, True ]
     CLS_TOKEN: [ False, False, False ]
     POS_EMBED: [ False, False, False ]
@@ -63,20 +64,21 @@ MODEL:
     PADDING_Q: [ 1, 1, 1 ]
     STRIDE_Q: [ 1, 1, 1 ]
     FREEZE_BN: true
-  HEAD_TYPE: CORNER #CORNER
-  HIDDEN_DIM: 384
+  HEAD_TYPE: CORNER
+  HIDDEN_DIM: 1024
   NUM_OBJECT_QUERIES: 1
   POSITION_EMBEDDING: sine
   PREDICT_MASK: false
 TRAIN:
-  BACKBONE_MULTIPLIER: 1.0
-  BATCH_SIZE: 8
+  BACKBONE_MULTIPLIER: 0.1
+  BATCH_SIZE: 16  # 16 for tesla V100(32 G)
   DEEP_SUPERVISION: false
   EPOCH: 500
-  GIOU_WEIGHT: 2.0
+  FREEZE_STAGE0: true
+  IOU_WEIGHT: 2.0
   GRAD_CLIP_NORM: 0.1
   L1_WEIGHT: 5.0
-  LR: 0.0001 #0.0001
+  LR: 0.0001
   LR_DROP_EPOCH: 400
   NUM_WORKER: 8
   OPTIMIZER: ADAMW
