+++ experiments/mixformer_cvt_online/baseline.yaml
@@ -40,6 +40,7 @@ DATA:
     - 1
     SAMPLE_PER_EPOCH: 10000
 MODEL:
+  VIT_TYPE: 'cvt_21'
   BACKBONE:
     PRETRAINED: true
     PRETRAINED_PATH: ''
@@ -77,14 +78,14 @@ TRAIN:
   BACKBONE_MULTIPLIER: 0.1
   BATCH_SIZE: 32
   DEEP_SUPERVISION: false
-  EPOCH: 40
+  EPOCH: 30 # 40
   TRAIN_SCORE: true
-  GIOU_WEIGHT: 2.0
+  IOU_WEIGHT: 2.0
   GRAD_CLIP_NORM: 0.1
   L1_WEIGHT: 5.0
   SCORE_WEIGHT: 1.0
   LR: 0.0001
-  LR_DROP_EPOCH: 30
+  LR_DROP_EPOCH: 20
   NUM_WORKER: 8
   OPTIMIZER: ADAMW
   PRINT_INTERVAL: 50
@@ -105,13 +106,13 @@ TEST:
     OTB: [6]
     GOT10K_TEST: [10]
     TRACKINGNET: [25]
-    VOT20: [7]
-    VOT20LT: [200]
+    VOT20: [10]
+#    VOT20LT: [200]
   ONLINE_SIZES:
-    LASOT: [3]
-    UAV: [3]
-    OTB: [2]
-    GOT10K_TEST: [3]
-    TRACKINGNET: [3]
+    LASOT: [2]
+    UAV: [1]
+    OTB: [3]
+    GOT10K_TEST: [2]
+    TRACKINGNET: [1]
     VOT20: [5]
-    VOT20LT: [3]
+#    VOT20LT: [3]
