+++ experiments/mixformer_cvt_online/baseline_large.yaml
@@ -40,21 +40,22 @@ DATA:
     - 1
     SAMPLE_PER_EPOCH: 10000
 MODEL:
+  VIT_TYPE: 'cvt_24'
   BACKBONE:
-    PRETRAINED: true
+    PRETRAINED: false
     PRETRAINED_PATH: ''
     INIT: 'trunc_norm'
     NUM_STAGES: 3
     PATCH_SIZE: [ 7, 3, 3 ]
     PATCH_STRIDE: [ 4, 2, 2 ]
     PATCH_PADDING: [ 2, 1, 1 ]
-    DIM_EMBED: [ 64, 192, 384 ]
-    NUM_HEADS: [ 1, 3, 6 ]
-    DEPTH: [ 1, 4, 16 ]
+    DIM_EMBED: [ 192, 768, 1024 ]
+    NUM_HEADS: [ 3, 12, 16 ]
+    DEPTH: [ 2, 2, 12 ]
     MLP_RATIO: [ 4.0, 4.0, 4.0 ]
     ATTN_DROP_RATE: [ 0.0, 0.0, 0.0 ]
     DROP_RATE: [ 0.0, 0.0, 0.0 ]
-    DROP_PATH_RATE: [ 0.0, 0.0, 0.1 ]
+    DROP_PATH_RATE: [ 0.0, 0.0, 0.3 ]
     QKV_BIAS: [ True, True, True ]
     CLS_TOKEN: [ False, False, False ]
     POS_EMBED: [ False, False, False ]
@@ -67,24 +68,25 @@ MODEL:
     FREEZE_BN: true
   HEAD_FREEZE_BN: True
   PRETRAINED_STAGE1: True
+#  STAGE1_MODEL: '/data/cyt/experiments/cvtt/models/TSP_Online_cvtw16_22k_ep0475.pth.tar'
   HEAD_TYPE: CORNER
-  HIDDEN_DIM: 384
+  HIDDEN_DIM: 1024
   NLAYER_HEAD: 3
   NUM_OBJECT_QUERIES: 1
   POSITION_EMBEDDING: sine
   PREDICT_MASK: false
 TRAIN:
   BACKBONE_MULTIPLIER: 0.1
-  BATCH_SIZE: 32
+  BATCH_SIZE: 16
   DEEP_SUPERVISION: false
-  EPOCH: 30 # 40
+  EPOCH: 40
   TRAIN_SCORE: true
-  GIOU_WEIGHT: 2.0
+  IOU_WEIGHT: 2.0
   GRAD_CLIP_NORM: 0.1
   L1_WEIGHT: 5.0
   SCORE_WEIGHT: 1.0
   LR: 0.0001
-  LR_DROP_EPOCH: 20
+  LR_DROP_EPOCH: 30
   NUM_WORKER: 8
   OPTIMIZER: ADAMW
   PRINT_INTERVAL: 50
@@ -100,18 +102,18 @@ TEST:
   TEMPLATE_FACTOR: 2.0
   TEMPLATE_SIZE: 128
   UPDATE_INTERVALS:
-    LASOT: [200]
-    UAV: [200]
-    OTB: [6]
-    GOT10K_TEST: [10]
-    TRACKINGNET: [25]
-    VOT20: [10]
-#    VOT20LT: [200]
+    LASOT: [ 200 ]
+    UAV: [ 25 ]
+    OTB: [ 6 ]
+    GOT10K_TEST: [ 200 ]
+    TRACKINGNET: [ 25 ]
+    VOT20: [ 10 ]
+#    VOT20LT: [ 200 ]
   ONLINE_SIZES:
-    LASOT: [2]
-    UAV: [1]
-    OTB: [3]
-    GOT10K_TEST: [2]
-    TRACKINGNET: [1]
-    VOT20: [5]
-#    VOT20LT: [3]
+    LASOT: [ 2 ]
+    UAV: [ 3 ]
+    OTB: [ 2 ]
+    GOT10K_TEST: [ 2 ]
+    TRACKINGNET: [ 3 ]
+    VOT20: [ 5 ]
+#    VOT20LT: [ 3 ]
