+++ experiments/mixformer_vit_online/baseline_large.yaml
@@ -81,12 +81,14 @@ TEST:
     LASOT: [200]
     GOT10K_TEST: [200]
     TRACKINGNET: [25]
+    UAV: [25]
+    OTB: [12]
     VOT20: [10]
     VOT20LT: [200]
   ONLINE_SIZES:
     LASOT: [2]
-    UAV: [3]
-    OTB: [2]
+    UAV: [2]
+    OTB: [5]
     GOT10K_TEST: [2]
-    TRACKINGNET: [3]
+    TRACKINGNET: [2]
     VOT20: [5]
