+++ lib/config/mixformer_cvt/config.py
@@ -2,7 +2,7 @@ from easydict import EasyDict as edict
 import yaml
 
 """
-Add default config for MixFormerOnline.
+Add default config for MixFormer.
 """
 cfg = edict()
 
@@ -13,6 +13,7 @@ cfg.MODEL.HIDDEN_DIM = 384
 cfg.MODEL.NUM_OBJECT_QUERIES = 1
 cfg.MODEL.POSITION_EMBEDDING = 'sine'  # sine or learned
 cfg.MODEL.PREDICT_MASK = False
+cfg.MODEL.VIT_TYPE = 'cvt_21'
 # MODEL.BACKBONE
 cfg.MODEL.BACKBONE = edict()
 cfg.MODEL.BACKBONE.PRETRAINED = True
@@ -39,14 +40,10 @@ cfg.MODEL.BACKBONE.STRIDE_KV = [ 2, 2, 2 ]
 cfg.MODEL.BACKBONE.PADDING_Q = [ 1, 1, 1 ]
 cfg.MODEL.BACKBONE.STRIDE_Q = [ 1, 1, 1 ]
 cfg.MODEL.BACKBONE.FREEZE_BN = True
-cfg.MODEL.PRETRAINED_STAGE1 = True
-cfg.MODEL.NLAYER_HEAD = 3
-cfg.MODEL.HEAD_FREEZE_BN = False
+
 
 # TRAIN
 cfg.TRAIN = edict()
-cfg.TRAIN.TRAIN_SCORE = True
-cfg.TRAIN.SCORE_WEIGHT = 1.0
 cfg.TRAIN.LR = 0.0001
 cfg.TRAIN.WEIGHT_DECAY = 0.0001
 cfg.TRAIN.EPOCH = 500
@@ -55,7 +52,7 @@ cfg.TRAIN.BATCH_SIZE = 16
 cfg.TRAIN.NUM_WORKER = 8
 cfg.TRAIN.OPTIMIZER = "ADAMW"
 cfg.TRAIN.BACKBONE_MULTIPLIER = 0.1
-cfg.TRAIN.GIOU_WEIGHT = 2.0
+cfg.TRAIN.IOU_WEIGHT = 2.0
 cfg.TRAIN.L1_WEIGHT = 5.0
 cfg.TRAIN.DEEP_SUPERVISION = False
 cfg.TRAIN.FREEZE_STAGE0 = False
@@ -69,7 +66,6 @@ cfg.TRAIN.SCHEDULER.DECAY_RATE = 0.1
 
 # DATA
 cfg.DATA = edict()
-cfg.DATA.SAMPLER_MODE = 'trident_pro'
 cfg.DATA.MEAN = [0.485, 0.456, 0.406]
 cfg.DATA.STD = [0.229, 0.224, 0.225]
 cfg.DATA.MAX_SAMPLE_INTERVAL = 200
@@ -110,17 +106,6 @@ cfg.TEST.UPDATE_INTERVALS.GOT10K_TEST = [200]
 cfg.TEST.UPDATE_INTERVALS.TRACKINGNET = [200]
 cfg.TEST.UPDATE_INTERVALS.VOT20 = [200]
 cfg.TEST.UPDATE_INTERVALS.VOT20LT = [200]
-cfg.TEST.UPDATE_INTERVALS.OTB = [200]
-cfg.TEST.UPDATE_INTERVALS.UAV = [200]
-
-cfg.TEST.ONLINE_SIZES= edict()
-cfg.TEST.ONLINE_SIZES.LASOT = [3]
-cfg.TEST.ONLINE_SIZES.GOT10K_TEST = [3]
-cfg.TEST.ONLINE_SIZES.TRACKINGNET = [3]
-cfg.TEST.ONLINE_SIZES.VOT20 = [3]
-cfg.TEST.ONLINE_SIZES.VOT20LT = [3]
-cfg.TEST.ONLINE_SIZES.OTB = [3]
-cfg.TEST.ONLINE_SIZES.UAV = [3]
 
 
 def _edict2dict(dest_dict, src_edict):
