--- lib/models/mixformer/__init__.py
@@ -1,2 +1,0 @@
-from .mixformer import build_mixformer
-from .mixformer_online import build_mixformer_online_score
