+++ lib/models/mixformer_convmae/mixformer.py
@@ -1,32 +1,16 @@
-import math
 from functools import partial
-import warnings
 
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-
+from einops import rearrange
 from timm.models.layers import DropPath, Mlp, trunc_normal_
 
 from lib.utils.misc import is_main_process
-from lib.models.mixformer.head import build_box_head
+from lib.models.mixformer_cvt.head import build_box_head
+from lib.models.mixformer_cvt.utils import to_2tuple
 from lib.utils.box_ops import box_xyxy_to_cxcywh, box_cxcywh_to_xyxy
 from lib.models.mixformer_vit.pos_utils import get_2d_sincos_pos_embed
-
-from einops import rearrange
-from itertools import repeat
-import collections.abc
-
-
-# From PyTorch internals
-def _ntuple(n):
-    def parse(x):
-        if isinstance(x, collections.abc.Iterable):
-            return x
-        return tuple(repeat(x, n))
-    return parse
-
-to_2tuple = _ntuple(2)
 
 
 class CMlp(nn.Module):
