+++ lib/models/mixformer_cvt/utils.py
@@ -3,7 +3,19 @@ Mixformer utils.
 """
 
 import torch
+from itertools import repeat
+import collections.abc
 
+
+# From PyTorch internals
+def _ntuple(n):
+    def parse(x):
+        if isinstance(x, collections.abc.Iterable):
+            return x
+        return tuple(repeat(x, n))
+    return parse
+
+to_2tuple = _ntuple(2)
 
 
 class FrozenBatchNorm2d(torch.nn.Module):
