+++ lib/test/parameter/mixformer_cvt.py
@@ -1,7 +1,7 @@
 from lib.test.utils import TrackerParams
 import os
 from lib.test.evaluation.environment import env_settings
-from lib.config.mixformer.config import cfg, update_config_from_file
+from lib.config.mixformer_cvt.config import cfg, update_config_from_file
 
 
 def parameters(yaml_name: str, search_area_scale=None):
@@ -9,7 +9,7 @@ def parameters(yaml_name: str, search_area_scale=None):
     prj_dir = env_settings().prj_dir
     save_dir = env_settings().save_dir
     # update default config from yaml file
-    yaml_file = os.path.join(prj_dir, 'experiments/mixformer/%s.yaml' % yaml_name)
+    yaml_file = os.path.join(prj_dir, 'experiments/mixformer_cvt/%s.yaml' % yaml_name)
     update_config_from_file(yaml_file)
     params.cfg = cfg
     print("test config: ", cfg)
