+++ lib/train/train_script_mixformer.py
@@ -10,7 +10,7 @@ from torch.nn.parallel import DistributedDataParallel as DDP
 # some more advanced functions
 from .base_functions import *
 # network related
-from lib.models.mixformer import build_mixformer, build_mixformer_online_score
+from lib.models.mixformer_cvt import build_mixformer_cvt, build_mixformer_cvt_online_score
 from lib.models.mixformer_vit import build_mixformer_vit, build_mixformer_vit_online_score
 from lib.models.mixformer_convmae import build_mixformer_convmae, build_mixformer_convmae_online_score
 # forward propagation related
@@ -55,10 +55,10 @@ def run(settings):
     loader_train, loader_val = build_dataloaders(cfg, settings)
 
     # Create network
-    if settings.script_name == "mixformer":
-        net = build_mixformer(cfg)
-    elif settings.script_name == "mixformer_online":
-        net = build_mixformer_online_score(cfg, settings)
+    if settings.script_name == "mixformer_cvt":
+        net = build_mixformer_cvt(cfg)
+    elif settings.script_name == "mixformer_online_cvt":
+        net = build_mixformer_cvt_online_score(cfg, settings)
     elif settings.script_name == "mixformer_vit":
         net = build_mixformer_vit(cfg)
     elif settings.script_name == "mixformer_vit_online":
@@ -82,11 +82,11 @@ def run(settings):
     settings.distill_loss_type = getattr(cfg.TRAIN, "DISTILL_LOSS_TYPE", "KL")
     # settings.save_every_epoch = True
     # Loss functions and Actors
-    if settings.script_name in ["mixformer", "mixformer_vit", "mixformer_convmae"]:
+    if settings.script_name in ["mixformer_cvt", "mixformer_vit", "mixformer_convmae"]:
         objective = {'ciou': ciou_loss, 'l1': l1_loss}
         loss_weight = {'ciou': cfg.TRAIN.IOU_WEIGHT, 'l1': cfg.TRAIN.L1_WEIGHT}
         actor = MixFormerActor(net=net, objective=objective, loss_weight=loss_weight, settings=settings)
-    elif settings.script_name in ["mixformer_online", "mixformer_vit_online", "mixformer_convmae_online"]:
+    elif settings.script_name in ["mixformer_cvt_online", "mixformer_vit_online", "mixformer_convmae_online"]:
         objective = {'ciou': ciou_loss, 'l1': l1_loss, 'score': BCEWithLogitsLoss()}
         loss_weight = {'ciou': cfg.TRAIN.IOU_WEIGHT, 'l1': cfg.TRAIN.L1_WEIGHT, 'score': cfg.TRAIN.SCORE_WEIGHT}
         actor = MixFormerActor(net=net, objective=objective, loss_weight=loss_weight, settings=settings, run_score_head=True)
