+++ lib/utils/box_ops.py
@@ -1,3 +1,4 @@
+import math
 import torch
 from torchvision.ops.boxes import box_area
 import numpy as np
@@ -93,6 +94,62 @@ def giou_loss(boxes1, boxes2):
     """
     giou, iou = generalized_box_iou(boxes1, boxes2)
     return (1 - giou).mean(), iou
+
+
+
+def ciou_loss(bboxes1, bboxes2):
+    """
+    :param boxes1: (N, 4) (x1,y1,x2,y2)
+    :param boxes2: (N, 4) (x1,y1,x2,y2)
+    :return:
+    """
+    rows = bboxes1.shape[0]
+    cols = bboxes2.shape[0]
+    cious = torch.zeros((rows, cols))
+    if rows * cols == 0:
+        return cious
+    exchange = False
+    if bboxes1.shape[0] > bboxes2.shape[0]:
+        bboxes1, bboxes2 = bboxes2, bboxes1
+        cious = torch.zeros((cols, rows))
+        exchange = True
+    w1 = bboxes1[:, 2] - bboxes1[:, 0]
+    h1 = bboxes1[:, 3] - bboxes1[:, 1]
+    w2 = bboxes2[:, 2] - bboxes2[:, 0]
+    h2 = bboxes2[:, 3] - bboxes2[:, 1]
+    area1 = w1 * h1
+    area2 = w2 * h2
+    center_x1 = (bboxes1[:, 0] + bboxes1[:, 2]) / 2.
+    center_y1 = (bboxes1[:, 1] + bboxes1[:, 3]) / 2.
+    center_x2 = (bboxes2[:, 0] + bboxes2[:, 2]) / 2.
+    center_y2 = (bboxes2[:, 1] + bboxes2[:, 3]) / 2.
+
+    inter_l = torch.max(center_x1 - w1 / 2,center_x2 - w2 / 2)
+    inter_r = torch.min(center_x1 + w1 / 2,center_x2 + w2 / 2)
+    inter_t = torch.max(center_y1 - h1 / 2,center_y2 - h2 / 2)
+    inter_b = torch.min(center_y1 + h1 / 2,center_y2 + h2 / 2)
+    inter_area = torch.clamp((inter_r - inter_l),min=0) * torch.clamp((inter_b - inter_t),min=0)
+
+    c_l = torch.min(center_x1 - w1 / 2,center_x2 - w2 / 2)
+    c_r = torch.max(center_x1 + w1 / 2,center_x2 + w2 / 2)
+    c_t = torch.min(center_y1 - h1 / 2,center_y2 - h2 / 2)
+    c_b = torch.max(center_y1 + h1 / 2,center_y2 + h2 / 2)
+
+    inter_diag = (center_x2 - center_x1)**2 + (center_y2 - center_y1)**2
+    c_diag = torch.clamp((c_r - c_l),min=0)**2 + torch.clamp((c_b - c_t),min=0)**2
+
+    union = area1+area2-inter_area
+    u = (inter_diag) / c_diag
+    iou = inter_area / union
+    v = (4 / (math.pi ** 2)) * torch.pow((torch.atan(w2 / h2) - torch.atan(w1 / h1)), 2)
+    with torch.no_grad():
+        S = (iou>0.5).float()
+        alpha= S*v/(1-iou+v)
+    cious = iou - u - alpha * v
+    cious = torch.clamp(cious,min=-1.0,max = 1.0)
+    if exchange:
+        cious = cious.T
+    return torch.mean(1-cious), iou
 
 
 def clip_box(box: list, H, W, margin=0):
